Nicolaskapfer/fix sshagent (#2)
* Fixing issues with SSH Agent

* Fixing issues in the first version

* Test install of wingetpathupdater

* Fixing issues with the SSH-Agent

* deleting Key variable

* Install modules for all users

* Signed Script

* Removed ExecutionPolicy Bypass (Code signed)

---------

+++ install-paw.ps1
@@ -37,10 +37,15 @@ Write-Host "Installing required softwares:" -ForegroundColor $Colors.Step
 
 ### WingetPathUpdater
 Write-Host " - Installing WingetPathUpdater..." -ForegroundColor $Colors.SubStep
-if(!$WhatIfPreference){
-    winget install jazzdelightsme.WingetPathUpdater --nowarn -h --accept-package-agreements --accept-source-agreements
+if(Test-Path "C:\Windows\System32\winget.ps1"){
+    Write-Host "[Skipped]" -ForegroundColor $Colors.Skipped
+}else{
+    if(!$WhatIfPreference){
+        winget install jazzdelightsme.WingetPathUpdater --nowarn -h --accept-package-agreements --accept-source-agreements
+    }
+    Write-Host "[Installed]" -ForegroundColor $Colors.Success
 }
-Write-Host "[Installed]" -ForegroundColor $Colors.Success
+
 ### Git installation
 Write-Host " - Installing Git... " -NoNewline -ForegroundColor $Colors.SubStep
 if(!(Get-Command "git" -ErrorAction SilentlyContinue).Path){
@@ -55,7 +60,7 @@ if(!(Get-Command "git" -ErrorAction SilentlyContinue).Path){
 
 ### OpenSSH
 Write-Host " - Installing OpenSSH..." -NoNewline -ForegroundColor $Colors.SubStep
-if(!(Get-Command "ssh-keygen" -ErrorAction SilentlyContinue).Path){
+if(!(Get-Command "ssh" -ErrorAction SilentlyContinue).Path){
     if(!$WhatIfPreference){
         Write-Host ""
         winget install Microsoft.OpenSSH.Beta --disable-interactivity --nowarn -h --accept-package-agreements --accept-source-agreements
@@ -65,7 +70,7 @@ if(!(Get-Command "ssh-keygen" -ErrorAction SilentlyContinue).Path){
     Write-Host "[Skipped]" -ForegroundColor $Colors.Skipped
 }
 
-Write-Host " - Setting up Git sshcommand..." -NoNewline -ForegroundColor $Colors.SubStep
+Write-Host " - Setting up Git sshCommand..." -NoNewline -ForegroundColor $Colors.SubStep
 git config --global core.sshCommand 'C:/Windows/System32/OpenSSH/ssh.exe'
 Write-Host "[OK]" -ForegroundColor $Colors.Success
 
@@ -119,20 +124,14 @@ if(!$WhatIfPreference){
 }
 Write-Host "[OK]" -ForegroundColor $Colors.Success
 
-Write-Host " - Installing SSH-Agent..." -NoNewline -ForegroundColor $Colors.SubStep
+Write-Host " - Installing and starting SSH-Agent..." -NoNewline -ForegroundColor $Colors.SubStep
 if(!$WhatIfPreference){
     if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)) {
         Start-Process PowerShell -Verb RunAs "-NoProfile -ExecutionPolicy Bypass -Command `"Get-Service ssh-agent | Set-Service -StartupType Automatic -PassThru | Start-Service`"";
     }
 }
 Write-Host "[OK]" -ForegroundColor $Colors.Success
-
-Write-Host " - Starting SSH-Agent..." -NoNewline -ForegroundColor $Colors.SubStep
-if(!$WhatIfPreference){
-    start ssh-agent
-    start-ssh-agent.cmd
-}
-Write-Host "[OK]" -ForegroundColor $Colors.Success
+$ENV:SSH_AUTH_SOCK = $null
 
 Write-Host " - Adding SSH key to the ssh-agent..." -ForegroundColor $Colors.SubStep
 ssh-add $Key
@@ -141,11 +140,12 @@ Write-Host "[OK]" -ForegroundColor $Colors.Success
 Write-Host "---------- SSH Key --------------"
 cat "$($Key).pub"
 Write-Host "---------- End ------------------"
+
+Remove-Variable -Name Key
+
 Write-Host " - Adding the SSH Key to the Github account..." -NoNewline -ForegroundColor $Colors.SubStep
 Start-Process "https://github.com/settings/ssh/new"
 Read-Host -Prompt "Add your public SSH key in your github profile. (Do not forget to authorize for SSO!) Once it's done, press any key to continue"
-
-Remove-Variable -Name Key
 
 ### Creating the necessary folders
 Write-Host " "
@@ -163,13 +163,17 @@ Foreach($path in $paths){
 }
 
 
+
+
 ### Installing necessary modules
 Write-Host " "
 Write-Host "Installing all required PowerShell modules:" -ForegroundColor $Colors.Step
 Write-Host -NoNewline " - Installing PowershellGet... " -ForegroundColor $Colors.SubStep
 if((Get-InstalledModule -Name PowershellGet -ErrorAction SilentlyContinue).count -eq 0){
     if(!$WhatIfPreference){
-        Install-Module PowershellGet -Force -Confirm:$false
+        if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)) {
+            Start-Process PowerShell -Verb RunAs "-NoProfile -ExecutionPolicy Bypass -Command `"Install-Module PowershellGet -Force -Confirm:$false`"";
+        }
     }
     Write-Host "[Installed]" -ForegroundColor $Colors.Success
 }else{
@@ -181,7 +185,9 @@ foreach($module in $msModules){
     Write-Host -NoNewline " - Installing $module... " -ForegroundColor $Colors.SubStep
     if((Get-InstalledModule -Name $module -ErrorAction SilentlyContinue).count -eq 0){
         if(!$WhatIfPreference){
-            Install-Module $module -Force -Confirm:$false
+            if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)) {
+                Start-Process PowerShell -Verb RunAs "-NoProfile -ExecutionPolicy Bypass -Command `"Install-Module $module -Force -Confirm:$false`"";
+            }
         }
         Write-Host "[Installed]" -ForegroundColor $Colors.Success
     }else{
@@ -244,3 +250,38 @@ Write-Host "|" -NoNewline -ForegroundColor $Colors.Frame
 Write-Host " Please restart your PowerShell 7 session to apply the changes. " -NoNewline -ForegroundColor $Colors.Text
 Write-Host "|" -ForegroundColor $Colors.Frame
 Write-Host "+----------------------------------------------------------------+" -ForegroundColor $Colors.Frame
+# SIG # Begin signature block
+# MIIGEQYJKoZIhvcNAQcCoIIGAjCCBf4CAQExDzANBglghkgBZQMEAgEFADB5Bgor
+# BgEEAYI3AgEEoGswaTA0BgorBgEEAYI3AgEeMCYCAwEAAAQQH8w7YFlLCE63JNLG
+# KX7zUQIBAAIBAAIBAAIBAAIBADAxMA0GCWCGSAFlAwQCAQUABCDgosxGjbut0zNX
+# Fia3hJgG5eBnP/co9jlNxfe2iS8OE6CCA3cwggNzMIIC+qADAgECAg9UAu4AoWdX
+# Oi5N4fIeXUMwCgYIKoZIzj0EAwQwKzEpMCcGA1UEAwwgRW1tYS1TbGVlcCBDZXJ0
+# aWZpY2F0ZSBBdXRob3JpdHkwHhcNMjQwODA2MTYwNzEzWhcNMjYxMTA5MTYwNzEz
+# WjCBqDELMAkGA1UEBhMCREUxDzANBgNVBAgMBkhlc3NlbjEaMBgGA1UEBwwRRnJh
+# bmtmdXJ0IGFtIE1haW4xEzARBgNVBAoMCkVtbWEtU2xlZXAxEDAOBgNVBAsMB0Vt
+# bWEgSVQxFzAVBgNVBAMMDk5pY29sYXMgS2FwZmVyMSwwKgYJKoZIhvcNAQkBFh1u
+# aWNvbGFzLmthcGZlckBlbW1hLXNsZWVwLmNvbTCCASIwDQYJKoZIhvcNAQEBBQAD
+# ggEPADCCAQoCggEBAO1Nb3oyoRoJOar6z0Gi2eOyV3kjPrcW7nTnfm5vQbzrGVwr
+# HpbmYkL7TeYz+rMYBEbRRhFK9wF5GT/M+S5dvEW8Ufq2aXm41IeZnx2mq5r48Vk4
+# /XJLPWBMALh08Tnpfhwq67kgMIzL2pN6wHP7e+l20eCOrW6klN3V8bfMtz6luhuG
+# wkLy1IYc0yyu72qNtaf/g3ScCQ5X9jQbbMgijL2s1SASx9/3UKnRhOSd3QWwzisx
+# ejc5E6rjeHHaPTHLxPzp5uUIivhhqOdLb4BpYIuQsGCNp5nB/7lXG4i3KMJIDbEv
+# fef9rQYgCbEHLEXpYcfzrvjujNCC2P7FlxozlCUCAwEAAaOBtzCBtDAJBgNVHRME
+# AjAAMB0GA1UdDgQWBBRNEEVRJubH944NocFeUZJHwR1ymjBmBgNVHSMEXzBdgBTS
+# M4pM4h6p9UrQNppF1Of9AAKK5aEvpC0wKzEpMCcGA1UEAwwgRW1tYS1TbGVlcCBD
+# ZXJ0aWZpY2F0ZSBBdXRob3JpdHmCFCYNGw7TIMaYx4jnPAoA52DDJaymMBMGA1Ud
+# JQQMMAoGCCsGAQUFBwMDMAsGA1UdDwQEAwIHgDAKBggqhkjOPQQDBANnADBkAjBk
+# TFR4vNInwGzEOfprM5V9lo8CqJ5q4yy0QGL+EDrQP3F3DCV0CmYutHH45olZU50C
+# MHpT92rcQC/4CfpY5StYSf6Y+BBSUXuyzKp889+IbZGcMAM1u6lBe9ozHMi7JM3T
+# qzGCAfAwggHsAgEBMD4wKzEpMCcGA1UEAwwgRW1tYS1TbGVlcCBDZXJ0aWZpY2F0
+# ZSBBdXRob3JpdHkCD1QC7gChZ1c6Lk3h8h5dQzANBglghkgBZQMEAgEFAKCBhDAY
+# BgorBgEEAYI3AgEMMQowCKACgAChAoAAMBkGCSqGSIb3DQEJAzEMBgorBgEEAYI3
+# AgEEMBwGCisGAQQBgjcCAQsxDjAMBgorBgEEAYI3AgEVMC8GCSqGSIb3DQEJBDEi
+# BCDdXzKheQk33QhjHtcJiE6haB1q5Iq19Ae38kyAmqa88jANBgkqhkiG9w0BAQEF
+# AASCAQBCs/UVxbJclU/hLXrSBAankPyVCw5vV8Pt3fgsWTMiGTx+u3vhX2DpEuh3
+# /1fvhBreRirGWFNbr8mRL3ER7g4WGjYUJVmZZ7BDSZu13ABa/5mTonyv0k4gzfw2
+# rDWERcq2vTm/qJP6zu5iKg0PDQFmgUpPfswArIWjbiRX+jN9ews0WobvYMB3Op/1
+# 1UgZeokLrR6/TsKSnI4BQx3PrLO6T1VLmtKF2qcgY+Br+8HIC4P39nTTA++xLnum
+# lkcbFIVM9vCfgEsWqTgzTLA6BPYTKhlP1HaCmGPmV+cXX7+MRtLnFvmn8qZ7Gf8i
+# YbuCPQPunOM4PAtROL+6yH7AgCje
+# SIG # End signature block
